+++ app/src/main/java/com/example/alldayfit/calendar/CalendarActivity.kt
@@ -7,6 +7,7 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.alldayfit.R
+import com.example.alldayfit.calendar.adapter.CalendarViewOptionsAdapter
 import com.example.alldayfit.databinding.CalendarActivityBinding
